Restructure transport for energy interface
